+++ include/redscore/platform/archive_manager.h
@@ -10,7 +10,7 @@
 #include "redscore/platform/archive.h"
 
 class ArchiveManager: public Archive{
-    using load_archive_callback = std::function<std::pair<bool, uint32>(ArchiveManager &manager, uint32 hash)>;
+    using load_archive_callback = std::function<std::pair<bool, uint64>(ArchiveManager &manager, uint64 hash)>;
 
 public:
 
@@ -18,18 +18,18 @@ public:
         : m_load_archive(std::move(load_archive)) {
     }
 
-    [[nodiscard]] bool is_mounted(uint32 archive_hash) const;
+    [[nodiscard]] bool is_mounted(uint64 archive_hash) const;
 
     void mount(std::unique_ptr<Archive> archive);
 
-    void unmount(uint32 archive_hash);
+    void unmount(uint64 archive_hash);
     void unmount(std::string_view name);
 
-    [[nodiscard]] bool has_file(uint32 hash) override;
+    [[nodiscard]] bool has_file(uint64 hash) override;
     [[nodiscard]] bool has_file(std::string_view name) override;
 
 
-    std::unique_ptr<IO::File> get_file(uint32 hash) override;
+    std::unique_ptr<IO::File> get_file(uint64 hash) override;
     std::unique_ptr<IO::File> get_file(std::string_view name) override;
 
     void all_entries(std::vector<ArchiveEntry> &entries) const override;
@@ -38,7 +38,7 @@ public:
         return "Root";
     }
 
-    uint32 hash() override {
+    uint64 hash() override {
         return 0;
     }
 
@@ -50,16 +50,16 @@ public:
     ArchiveManager& operator=(ArchiveManager&&) noexcept = default;
 
 protected:
-    std::unordered_map<uint32, std::unique_ptr<Archive>> m_archives;
+    std::unordered_map<uint64, std::unique_ptr<Archive>> m_archives;
     load_archive_callback m_load_archive;
 
     static constexpr size_t MAX_DYNAMIC_MOUNTS = 16;
 
-    void touch_dynamic_mount(uint32 hash);
+    void touch_dynamic_mount(uint64 hash);
     void evict_dynamic_mounts();
-    void forget_dynamic_mount(uint32 hash);
-    std::deque<uint32> m_dynamic_mount_order;
-    std::unordered_set<uint32> m_dynamic_mount_set;
+    void forget_dynamic_mount(uint64 hash);
+    std::deque<uint64> m_dynamic_mount_order;
+    std::unordered_set<uint64> m_dynamic_mount_set;
 
 };
