+++ src/platform/archive_manager.cpp
@@ -11,7 +11,7 @@
 
 static constexpr auto hash_string = std::hash<std::string_view>{};
 
-bool ArchiveManager::is_mounted(const uint32 archive_hash) const {
+bool ArchiveManager::is_mounted(const uint64 archive_hash) const {
     for (const auto &archive: m_archives | std::views::values) {
         if (archive->hash() == archive_hash) return true;
     }
@@ -22,7 +22,7 @@ void ArchiveManager::mount(std::unique_ptr<Archive> archive) {
     m_archives.emplace(archive->hash(), std::move(archive));
 }
 
-void ArchiveManager::unmount(const uint32 archive_hash) {
+void ArchiveManager::unmount(const uint64 archive_hash) {
     forget_dynamic_mount(archive_hash);
     m_archives.erase(archive_hash);
 }
@@ -32,7 +32,7 @@ void ArchiveManager::unmount(const std::string_view name) {
     unmount(hash_string(name));
 }
 
-bool ArchiveManager::has_file(const uint32 hash) {
+bool ArchiveManager::has_file(const uint64 hash) {
     ZoneScoped;
     for (const auto &archive: m_archives | std::views::values) {
         if (archive->has_file(hash)) return true;
@@ -44,7 +44,7 @@ bool ArchiveManager::has_file(const std::string_view name) {
     return has_file(hash_string(name));
 }
 
-std::unique_ptr<IO::File> ArchiveManager::get_file(const uint32 hash) {
+std::unique_ptr<IO::File> ArchiveManager::get_file(const uint64 hash) {
     ZoneScoped
     // ensure_parent_loaded(hash);
 
@@ -68,7 +68,7 @@ void ArchiveManager::all_entries(std::vector<ArchiveEntry> &entries) const {
     }
 }
 
-void ArchiveManager::touch_dynamic_mount(const uint32 hash) {
+void ArchiveManager::touch_dynamic_mount(const uint64 hash) {
     if (!m_dynamic_mount_set.insert(hash).second) {
         const auto it = std::ranges::find(m_dynamic_mount_order, hash);
         if (it != m_dynamic_mount_order.end()) {
@@ -96,7 +96,7 @@ void ArchiveManager::evict_dynamic_mounts() {
     }
 }
 
-void ArchiveManager::forget_dynamic_mount(const uint32 hash) {
+void ArchiveManager::forget_dynamic_mount(const uint64 hash) {
     if (!m_dynamic_mount_set.erase(hash)) {
         return;
     }
